Scripts page -> SideNav layout; update SideNav type system
