+++ frontend/pages/ManageControlsPage/OSSettings/OSSettingsNavItems.tsx
@@ -1,19 +1,24 @@
+import { InjectedRouter } from "react-router";
+
 import PATHS from "router/paths";
 import { ISideNavItem } from "pages/admin/components/SideNav/SideNav";
-import { IMdmProfile } from "interfaces/mdm";
 
 import DiskEncryption from "./cards/DiskEncryption";
 import CustomSettings from "./cards/CustomSettings";
+import { ICustomSettingsProps } from "./cards/CustomSettings/CustomSettings";
+import { IDiskEncryptionProps } from "./cards/DiskEncryption/DiskEncryption";
 
-interface IOSSettingsCardProps {
-  currentTeamId?: number;
-  profiles?: IMdmProfile[];
-  onProfileUpload?: () => void;
-  onProfileDelete?: () => void;
+export interface IOSSettingsCommonProps {
+  currentTeamId: number;
+  router: InjectedRouter; // v3
+  /** handler that fires when a change occures on the section (e.g. disk encryption
+   * enabled, profile uploaded) */
+  onMutation: () => void;
 }
 
-// TODO: types
-const OS_SETTINGS_NAV_ITEMS: ISideNavItem<IOSSettingsCardProps | any>[] = [
+type IOSSettingsCardProps = IDiskEncryptionProps | ICustomSettingsProps;
+
+const OS_SETTINGS_NAV_ITEMS: ISideNavItem<IOSSettingsCardProps>[] = [
   {
     title: "Disk encryption",
     urlSection: "disk-encryption",
