+++ frontend/router/index.tsx
@@ -286,6 +286,7 @@ const routes = (
                 <Route path="os-settings/:section" component={OSSettings} />
                 <Route path="setup-experience" component={SetupExperience} />
                 <Route path="scripts" component={Scripts} />
+                <Route path="scripts/:section" component={Scripts} />
                 <Route
                   path="setup-experience/:section"
                   component={SetupExperience}
